requirements.txt
flask_socketio
Flask
Flask-Session
eventlet
python-dotenv
